Warn about missing symbols that have no workaround
The ‘cfsetspeed’ and ‘cfmakeraw’ functions are not standard functions.
Guile-termios provides workarounds for those. The other functions
*are* standard termios functions, and if they are missing without
warning, the code

  (tcgetattr (port->fdes tty) tm)

would evaluate to

  (#f (port->fdes tty) tm)

and therefore fail to evaluate further.

This patch makes the module complain at load-time.

With this "make test" would complain, and debugging the previous
issue on cygwin would have been quite a bit simpler. :-)

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>