Add access to errno
Most of the functions signal errors by returning an integer < 0.
The kind of error is communicated via ‘errno’. So in order to
be able to properly react to errors, access is kind of required.

I don't mind doing it in here. But in the long run, something
like this should be provided in a central place. Be it guile-core,
or be it some "guile-posix" library.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>