Add frontends module to implement commonly used termios patterns
I'm mostly using guile-termios for interfacing to serial ports of embedded
hardware. Those usually need to be run in 8N1 mode with varying symbol rates,
with 9600bd being a common denominator, which I chose it as a default. Example
use case:

    (termios-8n1 "/dev/ttyACM2" termios-B115200)

Internally, this uses the (termios with-exceptions) module, so the function
will throw appropriately if any of its operations fails.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>